requirements.txt
python-telegram-bot>20,<21
sqlalchemy>2.0,<2.1
selenium>4.0,<5.0
undetected-chromedriver>3,<4
